[refactor] FolderCollection hook 정리
